ticketvote: Implmn cmdStartRunoffSub.
